Removed recordfieldstrait, initialize fields in function to avoid static functions, adjusted datasource setting to be more flexible, moved getRecordDataFormatterSpecClass to FinnaRecordTrait
